[Apple Silicon] Port tt-metal to macOS arm64 + enable ttsim on MacBooks
Enable building and running tt-metal with ttsim (Tenstorrent functional
simulator) natively on macOS Apple Silicon (aarch64-darwin). This allows
kernel developers to iterate locally on MacBooks without dedicated
Tenstorrent hardware.

**Test results on M4 Pro with ttsim Blackhole:**
- 12/13 programming examples PASS
- 10/11 unit_tests_device PASS
- 8121 LLK tests PASS (matmul, reduce, datacopy, risc_compute)
- 2 known ttsim gaps: coord_to_tile(1,0), hugepage-based L1↔PCIe test

The single programming example failure (hello_world_datatypes) and the
L1ToPCIe test failure are ttsim ISA/feature gaps, not macOS porting issues.

---

**Build system (cmake/):**
- aarch64-apple-clang-toolchain.cmake (NEW): Apple Clang toolchain for arm64
- cmake/stubs/elf.h (NEW): minimal ELF type/macro definitions for macOS
  (macOS uses Mach-O so elf.h is not in the SDK)
- compilers.cmake: accept AppleClang (MATCHES "Clang" instead of STREQUAL);
  skip libc++ find on macOS (it ships with the SDK)
- flatbuffers.cmake: use $<TARGET_FILE:flatc> generator expression so the
  built flatc binary is found without PATH manipulation

**CMakeLists.txt (root):**
- Add cmake/stubs/ as a SYSTEM include dir on macOS (provides elf.h)
- Find hwloc headers and library with Homebrew prefix hints
- Make `numa` link conditional: $<$<NOT:$<PLATFORM_ID:Darwin>>:numa>
- third_party/CMakeLists.txt: guard NUMA check with if(NOT APPLE); add
  Homebrew hints to hwloc find; guard nanobind fetch with WITH_PYTHON_BINDINGS

**tt_metal/hw/CMakeLists.txt (SFPI cross-compiler):**
- Fix BSD sed: "1{s/…//;p}" → "1s/…//p" (macOS sed syntax)
- Add TT_SKIP_SFPI_VERSION_CHECK cmake option for builds with substitute
  toolchains (e.g. brew riscv64-elf-gcc while building real SFPI)
- Skip ldd/ELF dependency check on macOS (NOT APPLE guard)
- Replace auto-SFPI-build-on-configure with a helpful FATAL_ERROR pointing
  to the build script

**tt_metal/sfpi-info.sh:**
- Replace bash 4+ ${var^^} with tr '[:lower:]' '[:upper:]' (macOS ships
  bash 3.2 which lacks the ^^ operator)
- Replace GNU sed {commands;} with grep+cut (BSD sed incompatibility)

**tt_metal/llrt/hal/codegen/codegen.sh:**
- Replace `realpath --relative-to` (GNU coreutils) with a portable bash
  path-stripping approach

**Host runtime fixes (Linux-only APIs → macOS stubs/equivalents):**
- jit_build/jit_build_utils.cpp: use _NSGetEnviron() for ::environ on macOS
- impl/device/device_manager.cpp: guard <numa.h>; guard NUMA API calls
  (numa_available, numa_node_of_cpu) with #ifdef __linux__; guard
  cpu_set_t/pthread_setaffinity_np similarly
- impl/threading/thread_pool.cpp: guard <numa.h> and all NUMA/affinity APIs
- impl/debug/dprint_server.cpp: pthread_setname_np signature differs on macOS
  (takes only the name string, not a pthread_t argument)
- impl/debug/inspector/data.cpp: use _NSGetEnviron() for environ
- impl/memory_tracking/memory_stats_shm.cpp: use libproc proc_name() instead
  of reading /proc/<pid>/comm
- tt_metal/fabric/fabric_host_utils.hpp: define HOST_NAME_MAX fallback via
  sys/param.h MAXHOSTNAMELEN
- tt_metal/fabric/mesh_graph.cpp: remove const from unordered_map key type
  (libc++ stricter than libstdc++ about hash<const T>)
- tt_metal/distributed/d2h_socket.cpp: guard <immintrin.h> with
  #ifdef __x86_64__; provide _mm_clflush/_mm_lfence no-op stubs for ARM
- tt_metal/distributed/mesh_buffer.cpp, impl/per_core_allocation/mesh_buffer.cpp:
  add explicit return types to std::visit lambdas (libc++ requires uniform
  return type, libstdc++ is lenient)
- tt_metal/impl/allocator/algorithms/free_list_opt.cpp: cast DeviceAddr and
  size_t explicitly for std::max (different types on macOS 64-bit)
- tt_metal/impl/buffers/buffer_distribution_spec.cpp: explicit cast
  SmallVector<size_t> → SmallVector<uint64_t> (unsigned long vs unsigned
  long long on macOS)
- tt_metal/impl/program/program.cpp: bridge uint64_t↔size_t for hash_combine
- tt_stl/tt_stl/reflection.hpp: add hash_combine(uint64_t&, …) overload for
  macOS where uint64_t≠size_t (unsigned long long vs unsigned long)

**Programming examples — DeviceAddr narrowing (uint64_t → uint32_t):**
All examples that pass buffer addresses as SetRuntimeArgs() values needed
explicit static_cast<uint32_t>() wrappers. Apple Clang (libc++) treats
narrowing conversions in braced init-lists as errors; libstdc++ warns only.
Affected: add_2_integers_in_riscv, NoC_tile_transfer, loopback,
eltwise_binary, eltwise_sfpu, hello_world_datatypes_kernel,
sfpu_eltwise_chain, vecadd_multi_core, matmul_multi_core.

Also: xy_pair.x/y are size_t (unsigned long), not uint32_t, so needed
explicit casts in NoC tile transfer and test_device.cpp as well.

**tt_metal/third_party/umd:**
Updated submodule pointer to include macOS porting changes (separate commit
in the umd submodule).

---

**External dependencies needed (brew):**
cmake ninja bash gcc riscv64-elf-gcc hwloc bison flex gawk texinfo gmp mpfr
libmpc expat

**SFPI cross-compiler for macOS:**
Build from source using Homebrew GCC (not Apple Clang, which has libc++
conflicts with GCC's own headers). Key fixes to SFPI source:
- binutils/zlib/zutil.h: guard fdopen redefinition with !defined(__MACH__)
- gcc/gcc/config/riscv/tt/rvtt-protos.h: add missing SFPXOR_MOD1_USE_VB=0
- scripts/build.sh: nproc → nproc 2>/dev/null || sysctl -n hw.ncpu

**ttsim:**
Built from tenstorrent/ttsim-private — already has macOS/aarch64 support,
no changes needed. Output: libttsim_bh.so / libttsim_wh.so (Mach-O arm64).

+++ tests/tt_metal/tt_metal/device/test_device.cpp
@@ -424,7 +424,7 @@ TEST_F(MeshDeviceFixture, VerifyLogicalToVirtualMap) {
         tt_metal::DataMovementConfig{
             .processor = tt_metal::DataMovementProcessor::RISCV_0,
             .noc = tt_metal::NOC::NOC_0,
-            .compile_args = {kernel0_l1_address, logical_grid_size.x, logical_grid_size.y}});
+            .compile_args = {kernel0_l1_address, static_cast<uint32_t>(logical_grid_size.x), static_cast<uint32_t>(logical_grid_size.y)}});
 
     auto kernel1_l1_address = l1_unreserved_base + (tt::round_up(logical_grid_size.x, 4) * sizeof(uint32_t)) +
                               (tt::round_up(logical_grid_size.y, 4) * sizeof(uint32_t));
@@ -435,7 +435,7 @@ TEST_F(MeshDeviceFixture, VerifyLogicalToVirtualMap) {
         tt_metal::DataMovementConfig{
             .processor = tt_metal::DataMovementProcessor::RISCV_1,
             .noc = tt_metal::NOC::NOC_1,
-            .compile_args = {kernel1_l1_address, logical_grid_size.x, logical_grid_size.y}});
+            .compile_args = {static_cast<uint32_t>(kernel1_l1_address), static_cast<uint32_t>(logical_grid_size.x), static_cast<uint32_t>(logical_grid_size.y)}});
 
     distributed::EnqueueMeshWorkload(cq, workload, false);
